backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.14",
    "fpdf>=1.7.2",
    "motor>=3.7.1",
    "openai-agents>=0.1.0",
    "playwright>=1.53.0",
    "pydantic[email]>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
